[FileWatcher] Implement Inotify as a fallback alternative
The way capabilities and seccomp works in container environments are
... Interesting. By default fanotify_init syscall is blocked, so
fanotify is not able to start in a container unless one fucks around
with seccomp. Inotify is apparently not blocked by default, so instead
of requiering tinkering with seccomp to run kjor in a container,
default to inotify for now.