[SSE] Allow the server to notify the client itself that it has rebooted
This is a bit of a WiP:

Is it necessary to have ones own SSE if the server can notify it
self. Sort of? In the monitor process one can implement a timeout
forcing the browser to reload even if the server didn't post. But then
the question becomes, why didn't it post? If it just because it is not
implemented, thats fine. But if it is because the boot is longer than
the timeout, then it becomes a problem.